requirements.txt
anyio==4.2.0
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
editables==0.5
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
google-api-core==2.15.0
google-api-python-client==2.112.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
h11==0.14.0
hatch==1.9.1
hatchling==1.21.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.3.0
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
mysql-connector-python==8.2.0
numba==0.58.1
oauthlib==3.2.2
openai-whisper==20231117
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
playsound==1.3.0
pluggy==1.3.0
praw==7.7.1
prawcore==2.4.0
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
selenium==4.16.0
shellingham==1.5.4
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tiktoken==0.5.2
toml==0.10.2
tomli_w==1.0.0
tomlkit==0.12.3
torch==2.1.2
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
trove-classifiers==2023.11.29
typing_extensions==4.9.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.1.0
userpath==1.9.1
virtualenv==20.25.0
webdriver-manager==4.0.1
websocket-client==1.7.0
wsproto==1.2.0
zipp==3.17.0
zstandard==0.22.0
